pyproject.toml
[project]
name = "hospitopt"
version = "0.1.0"
description = "Solution to optimize hospital-patitent allocation."
readme = "README.md"
requires-python = ">=3.14"
authors = [{ name = "Daniel Sousa" }]
urls = { "source" = "https://github.com/dhcsousa/hospitopt" }
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "google-maps-routing>=0.8.0",
    "greenlet>=3.3.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyomo>=6.9.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.3",
    "ipykernel>=7.1.0",
    "mypy>=1.19.1",
    "nbstripout>=0.9.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]
vis = [
    "folium>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core", "src/workers", "src/api"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true

[tool.black]
line-length = 120

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.bandit]
exclude_dirs = ["scripts"]

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']
